+++ src/components/RegistrationForm/form-field/index.ts
@@ -1,16 +1,3 @@
 export { default as FormField } from './FormField';
-export type {
-  InputType,
-  FormFieldType,
-  FormFieldBaseOptions,
-  FormFieldOptions,
-  FormInputs,
-} from './types';
+export type { FormFieldType, FormFieldBaseOptions, FormFieldOptions } from './types';
 export { FormFieldOptionsContext } from './context';
-export {
-  type FieldName,
-  defautFormValues,
-  testFormValues,
-  formFieldsOptions,
-  formFields,
-} from './config';
