+++ src/components/RegistrationForm/form-field/types.ts
@@ -1,27 +1,5 @@
+import { FormInputs, InputType } from '../types';
 import { ValidationOptions } from '../validation/types';
-
-export type FormInputs = {
-  name: string;
-  password: string;
-  email: string;
-  birthday: string;
-  country: string;
-  programmingLanguage: string | string[];
-  gender: string;
-  avatar: FileList | File;
-  subscribe: string;
-};
-
-export type InputType =
-  | 'text'
-  | 'password'
-  | 'date'
-  | 'email'
-  | 'select'
-  | 'radio'
-  | 'checkbox'
-  | 'file'
-  | '';
 
 export type FormFieldType = 'list' | '';
